Refining from multiplayer test
- Added more messages into languages file
- Adjusted hologram texts for better readability
- Closing arena now removes player and spectator count in portal
- Adjusted vampire kit lore and ability
- Fixed bug where permissions applied to all commands
- Admin command changed from /vd to /vd admin
- Fixed issue where the arena could duplicate ending situations
- Fixed bug where people with giant kit would respawn without the kit benefit
- More robust error handling for commands
- Selecting a kit now closes the selection inventory
- Fixed bug where removing an arena wouldn't remove the arenaboards and caused the NPC to reappear
- Updated spawn tables to push back skeleton spawning

+++ src/me/theguyhere/villagerdefense/game/displays/ArenaBoard.java
@@ -34,13 +34,17 @@ public class ArenaBoard {
 	}
 
 	public void refreshArenaBoard(int arena) {
-		arenaBoards[arena].delete();
-		addHolo(utils.getConfigLocationNoPitch("arenaBoard." + arena), game.arenas.get(arena));
+		if (arenaBoards[arena] != null) {
+			arenaBoards[arena].delete();
+			addHolo(utils.getConfigLocationNoPitch("arenaBoard." + arena), game.arenas.get(arena));
+		}
 	}
 
 	public void removeArenaBoard(int arena) {
-		arenaBoards[arena].delete();
-		arenaBoards[arena] = null;
+		if (arenaBoards[arena] != null) {
+			arenaBoards[arena].delete();
+			arenaBoards[arena] = null;
+		}
 	}
 
 	public void addHolo(Location location, Arena arena) {
@@ -72,8 +76,8 @@ public class ArenaBoard {
 		info.add(Utils.format("&6&l" + arena.getName() + " Records"));
 		if (!arena.getSortedDescendingRecords().isEmpty())
 			arena.getSortedDescendingRecords().stream().forEachOrdered(record -> {
-				StringBuilder firstLine = new StringBuilder("&fWave &b" + record.getWave() + " &f-&7");
-				StringBuilder secondLine = new StringBuilder("&7");
+				StringBuilder firstLine = new StringBuilder("Wave &b" + record.getWave() + " &f-");
+				StringBuilder secondLine = new StringBuilder();
 				if (record.getPlayers().size() > 6) {
 					for (int i = 0; i < record.getPlayers().size() / 2; i++)
 						firstLine.append(" ").append(record.getPlayers().get(i)).append(",");
